#2 #8 Extract user properties to configuration metadata and global context

+++ project/src/main/java/net/piotrl/music/aggregation/repository/AggregationCrudRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface AggregationCrudRepository extends CrudRepository<AggregationEntity, Integer> {
+public interface AggregationCrudRepository extends CrudRepository<AggregationEntity, Long> {
     Optional<AggregationEntity> findOneByAccountIdAndType(long accountId, String type);
 }
